Set Up data mining and plotting
I created a script to look through the commits on the source files of a project and get the author name and date of the commit. This info is saved to a .csv file that can be read by my scatterplot file and will create a scatter plot of all the different commits on the source files.